Refactor vectorstore to use core.LocalID for vector IDs
- Updated the Store interface and its implementations to replace uint64 IDs with core.LocalID.
- Modified methods in the Store, WAL, and related components to accommodate the new ID type.
- Adjusted tests and benchmarks to reflect the changes in ID handling.
- Removed the NewWithCapacity function in favor of a simplified New function.
- Introduced a Close method for resource management in the Store.
- Enhanced the Compact method to maintain ID mappings correctly.
- Ensured compatibility with existing functionality while improving type safety and clarity.

+++ benchmark_test/isolated/utils.go
@@ -9,7 +9,7 @@ import (
 func toTestUtilResults(results []index.SearchResult) []testutil.SearchResult {
 	out := make([]testutil.SearchResult, len(results))
 	for i, r := range results {
-		out[i] = testutil.SearchResult{ID: r.ID, Distance: r.Distance}
+		out[i] = testutil.SearchResult{ID: uint64(r.ID), Distance: r.Distance}
 	}
 	return out
 }
